CHANGE(SDS): Multi-conscience activation + Conf updates to decrease detection times when a node is down
 ##### SUMMARY
Activation of the multi conscience, which allows to stop any node of the cluster (3 nodes cluster)

SDS configuration update to allow the Conscience service to be more quickly aware that a service is unavailable.
The goal is to reduce the detection time when a node is down.
Update of the namespace, conscience and conscience-agent configuration

Update the Travis tests to simulate the loss of a server

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION
Steps to activate Multi-conscience :
1- in inventory.yml :
	 - openio_conscience_multiple_enable: true
	 - add the 3 nodes in the conscience configuration
2- use the openio_namespace_overwrite option to allow the namespace update :
	`ansible-playbook -i inventory.yml main.yml -e "openio_maintenance_mode=false" -e "openio_namespace_overwrite=true"`